requirements.txt
supabase==2.15.0
python-dotenv==1.1.0
pymodbus==3.6.6
